+++ packages/mime-types/src/db.ts
