+++ packages/mime-types/src/index.ts
