+++ packages/mime-types/tsup.config.ts
@@ -4,7 +4,7 @@ import { config } from "@uploadthing/tsup-config";
 
 export default defineConfig((opts) => ({
   ...config,
-  entry: ["./index.ts", "./db.ts"],
+  entry: ["./src/index.ts", "./src/db.ts"],
   clean: !opts.watch,
   async onSuccess() {
     // void
