--- packages/react/hooks.ts
@@ -1,1 +1,0 @@
-export { generateReactHelpers } from "./src/useUploadThing";
