+++ packages/react/tsconfig.sourcemap.json
@@ -7,5 +7,5 @@
     "emitDeclarationOnly": true,
     "declarationDir": "dist"
   },
-  "exclude": ["tsup.config.ts"]
+  "exclude": ["tsup.config.ts", "test"]
 }
