+++ packages/uploadthing/src/internal/next/approuter.ts
@@ -1,10 +1,7 @@
 import { UPLOADTHING_VERSION } from "../../constants";
-import type { RouterWithConfig } from "../../internal/handler";
-import {
-  buildPermissionsInfoHandler,
-  buildRequestHandler,
-} from "../../internal/handler";
-import type { FileRouter } from "../../internal/types";
+import type { RouterWithConfig } from "../handler";
+import { buildPermissionsInfoHandler, buildRequestHandler } from "../handler";
+import type { FileRouter } from "../types";
 
 export const createNextRouteHandler = <TRouter extends FileRouter>(
   opts: RouterWithConfig<TRouter>,
