+++ packages/uploadthing/src/internal/next/pagerouter.ts
@@ -1,12 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { UPLOADTHING_VERSION } from "../../constants";
-import type { RouterWithConfig } from "../../internal/handler";
-import {
-  buildPermissionsInfoHandler,
-  buildRequestHandler,
-} from "../../internal/handler";
-import type { FileRouter } from "../../internal/types";
+import type { RouterWithConfig } from "../handler";
+import { buildPermissionsInfoHandler, buildRequestHandler } from "../handler";
+import type { FileRouter } from "../types";
 
 export const createNextPageApiHandler = <TRouter extends FileRouter>(
   opts: RouterWithConfig<TRouter>,
