+++ packages/uploadthing/src/next-legacy.ts
@@ -1,0 +1,6 @@
+import { createBuilder } from "./internal/upload-builder";
+
+export { createNextPageApiHandler } from "./internal/next/pagerouter";
+export type { FileRouter } from "./internal/types";
+
+export const createUploadthing = () => createBuilder<"pages">();
