+++ packages/uploadthing/src/next.ts
@@ -1,0 +1,6 @@
+import { createBuilder } from "./internal/upload-builder";
+
+export { createNextRouteHandler } from "./internal/next/approuter";
+export type { FileRouter } from "./internal/types";
+
+export const createUploadthing = () => createBuilder<"app">();
