+++ packages/uploadthing/src/server.ts
@@ -1,0 +1,6 @@
+import { createBuilder } from "./internal/upload-builder";
+
+export * from "./internal/types";
+export * as utapi from "./sdk";
+
+export const createUploadthing = () => createBuilder<"web">();
