+++ packages/uploadthing/test/upload-builder.test.ts
@@ -5,9 +5,8 @@ import type { NextRequest } from "next/server";
 import { expect, expectTypeOf, it } from "vitest";
 import { z } from "zod";
 
-import { genUploader } from "../client";
-import type { FileRouter, UnsetMarker } from "./internal/types";
-import { createBuilder } from "./upload-builder";
+import type { UnsetMarker } from "../src/internal/types";
+import { createBuilder } from "../src/internal/upload-builder";
 
 const badReqMock = {
   headers: {
@@ -178,20 +177,4 @@ it("smoke", async () => {
     input: { foo: "bar" },
   });
   expect(metadata).toEqual({ header1: "woohoo", userId: "123" });
-});
-
-it("genuploader", async () => {
-  const f = createBuilder();
-  const uploadable = f(["image", "video"]).onUploadComplete(() => {});
-
-  const router = { uploadable } satisfies FileRouter;
-
-  const uploader = genUploader<typeof router>();
-
-  try {
-    // @ts-expect-error - Argument of type '"random"' is not assignable to parameter of type '"uploadable"'
-    await uploader([], "random");
-  } catch (e) {
-    // expected this to error since we're not in a real env so it can't fetch
-  }
 });
