+++ packages/uploadthing/tsconfig.json
@@ -16,12 +16,6 @@
     "skipLibCheck": true,
     "strict": true
   },
-  "include": [
-    "src/**/*",
-    "*.ts",
-    "../shared/mime-types",
-    "../mime-types/db.ts",
-    "../shared/types.ts"
-  ],
+  "include": ["src/**/*", "*.ts", "test/**/*"],
   "exclude": ["node_modules"]
 }
