+++ packages/uploadthing/tsup.config.ts
@@ -4,6 +4,11 @@ import { config } from "@uploadthing/tsup-config";
 
 export default defineConfig((opts) => ({
   ...config,
-  entry: ["./client.ts", "./server.ts", "./next.ts", "./next-legacy.ts"],
+  entry: [
+    "./src/client.ts",
+    "./src/server.ts",
+    "./src/next.ts",
+    "./src/next-legacy.ts",
+  ],
   clean: !opts.watch,
 }));
